requirements.txt
alembic==1.3.1
blinker==1.4
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SimpleMDE==0.3.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gunicorn==20.0.2
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
markdown2==2.3.8
MarkupSafe==1.1.1
Pillow==6.2.1
psycopg2==2.8.4
python-dateutil==2.8.1
python-editor==1.0.4
six==1.13.0
SQLAlchemy==1.3.11
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
